api/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.2
matplotlib==3.9.2
python-dateutil==2.9.0.post0
plotly==5.18.0
kaleido==0.2.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
rapidfuzz==3.9.0
